design : 결과페이지 임시레이아웃 퍼블리싱
