+++ my-project/src/components/보류_reports copy/BehaviorReport.tsx
@@ -1,0 +1,5 @@
+const BehaviorReport = () => {
+  return <div>행동별 레포트 내용</div>;
+};
+
+export default BehaviorReport;
